+++ pkg/handler/ratelimit.go
@@ -1,4 +1,4 @@
-package service
+package handler
 
 import (
 	"fmt"
